+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Changelog
 
-### 0.5.1 (Next)
+### 0.6.0 (Next)
 
+* [#47](https://github.com/dblock/danger-changelog/pull/47): Configure `Config#placeholder_line` at plugin level - [@dblock](https://github.com/dblock).
+* [#47](https://github.com/dblock/danger-changelog/pull/47): Deprecate `Danger::Changelog.configure` - [@dblock](https://github.com/dblock).
+* [#47](https://github.com/dblock/danger-changelog/pull/47): Deprecate `check` in favor of `check!` - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.5.0 (2019/05/31)